+++ temba/utils/tests.py
@@ -10,7 +10,6 @@ from celery.app.task import Task
 from django_redis import get_redis_connection
 
 from django import forms
-from django.conf import settings
 from django.forms import ValidationError
 from django.template import Context, Template
 from django.test import TestCase, override_settings
@@ -39,7 +38,6 @@ from . import (
 )
 from .crons import clear_cron_stats, cron_task
 from .dates import date_range, datetime_to_str, datetime_to_timestamp, timestamp_to_datetime
-from .email import is_valid_address, send_simple_email
 from .fields import ExternalURLField, NameValidator
 from .templatetags.temba import short_datetime
 from .text import clean_string, decode_stream, generate_secret, generate_token, slugify_with, truncate, unsnakify
@@ -448,110 +446,6 @@ class TemplateTagTest(TembaTest):
             to_json(json.dumps({"special": '<script>alert("XSS");</script>'})),
             'JSON.parse("{\\u0022special\\u0022: \\u0022\\u003Cscript\\u003Ealert(\\u005C\\u0022XSS\\u005C\\u0022)\\u003B\\u003C/script\\u003E\\u0022}")',
         )
-
-
-class EmailTest(TembaTest):
-    @override_settings(SEND_EMAILS=True)
-    def test_send_simple_email(self):
-        send_simple_email(["recipient@bar.com"], "Test Subject", "Test Body")
-        self.assertOutbox(0, settings.DEFAULT_FROM_EMAIL, "Test Subject", "Test Body", ["recipient@bar.com"])
-
-        send_simple_email(["recipient@bar.com"], "Test Subject", "Test Body", from_email="no-reply@foo.com")
-        self.assertOutbox(1, "no-reply@foo.com", "Test Subject", "Test Body", ["recipient@bar.com"])
-
-    def test_is_valid_address(self):
-        valid_emails = [
-            # Cases from https://en.wikipedia.org/wiki/Email_address
-            "prettyandsimple@example.com",
-            "very.common@example.com",
-            "disposable.style.email.with+symbol@example.com",
-            "other.email-with-dash@example.com",
-            "x@example.com",
-            '"much.more unusual"@example.com',
-            '"very.unusual.@.unusual.com"@example.com'
-            '"very.(),:;<>[]".VERY."very@\\ "very".unusual"@strange.example.com',
-            "example-indeed@strange-example.com",
-            "#!$%&'*+-/=?^_`{}|~@example.org",
-            '"()<>[]:,;@\\"!#$%&\'-/=?^_`{}| ~.a"@example.org',
-            '" "@example.org',
-            "example@localhost",
-            "example@s.solutions",
-            # Cases from Django tests
-            "email@here.com",
-            "weirder-email@here.and.there.com",
-            "email@[127.0.0.1]",
-            "email@[2001:dB8::1]",
-            "email@[2001:dB8:0:0:0:0:0:1]",
-            "email@[::fffF:127.0.0.1]",
-            "example@valid-----hyphens.com",
-            "example@valid-with-hyphens.com",
-            "test@domain.with.idn.tld.उदाहरण.परीक्षा",
-            "email@localhost",
-            '"test@test"@example.com',
-            "example@atm.%s" % ("a" * 63),
-            "example@%s.atm" % ("a" * 63),
-            "example@%s.%s.atm" % ("a" * 63, "b" * 10),
-            '"\\\011"@here.com',
-            "a@%s.us" % ("a" * 63),
-        ]
-
-        invalid_emails = [
-            # Cases from https://en.wikipedia.org/wiki/Email_address
-            None,
-            "",
-            "abc",
-            "a@b",
-            " @ .c",
-            "a @b.c",
-            "{@flow.email}",
-            "Abc.example.com",
-            "A@b@c@example.com",
-            r'a"b(c)d,e:f;g<h>i[j\k]l@example.com'
-            'just"not"right@example.com'
-            'this is"not\allowed@example.com'
-            r'this\ still"not\\allowed@example.com'
-            "1234567890123456789012345678901234567890123456789012345678901234+x@example.com"
-            "john..doe@example.com"
-            "john.doe@example..com"
-            # Cases from Django tests
-            "example@atm.%s" % ("a" * 64),
-            "example@%s.atm.%s" % ("b" * 64, "a" * 63),
-            None,
-            "",
-            "abc",
-            "abc@",
-            "abc@bar",
-            "a @x.cz",
-            "abc@.com",
-            "something@@somewhere.com",
-            "email@127.0.0.1",
-            "email@[127.0.0.256]",
-            "email@[2001:db8::12345]",
-            "email@[2001:db8:0:0:0:0:1]",
-            "email@[::ffff:127.0.0.256]",
-            "example@invalid-.com",
-            "example@-invalid.com",
-            "example@invalid.com-",
-            "example@inv-.alid-.com",
-            "example@inv-.-alid.com",
-            'test@example.com\n\n<script src="x.js">',
-            # Quoted-string format (CR not allowed)
-            '"\\\012"@here.com',
-            "trailingdot@shouldfail.com.",
-            # Max length of domain name labels is 63 characters per RFC 1034.
-            "a@%s.us" % ("a" * 64),
-            # Trailing newlines in username or domain not allowed
-            "a@b.com\n",
-            "a\n@b.com",
-            '"test@test"\n@example.com',
-            "a@[127.0.0.1]\n",
-        ]
-
-        for email in valid_emails:
-            self.assertTrue(is_valid_address(email), "FAILED: %s should be a valid email" % email)
-
-        for email in invalid_emails:
-            self.assertFalse(is_valid_address(email), "FAILED: %s should be an invalid email" % email)
 
 
 class JsonTest(TembaTest):
